feat: fix StarExec operating system change

+++ src/main/java/de/uulm/sp/fmc/as4moco/data/SolverRunInstance.java
@@ -4,9 +4,10 @@ import de.uulm.sp.fmc.as4moco.solver.SolverResponse;
 import de.uulm.sp.fmc.as4moco.solver.SolverStatusEnum;
 import de.uulm.sp.fmc.as4moco.solver.SolverType;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
-public record SolverRunInstance(Optional<String> solver, SolverStatusEnum status, Optional<Double> solution, double runtime, SolverType solverType) {
+public record SolverRunInstance(Optional<String> solver, SolverStatusEnum status, Optional<BigDecimal> solution, double runtime, SolverType solverType) {
 
     public SolverRunInstance(SolverResponse solverResponse, double runtime) {
         this(solverResponse.solver(), solverResponse.status(), solverResponse.solution(), runtime, solverResponse.solverType());
