+++ src/main/java/de/uulm/sp/fmc/as4moco/solver/SolverResponse.java
@@ -1,8 +1,9 @@
 package de.uulm.sp.fmc.as4moco.solver;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
-public record SolverResponse(Optional<String> solver, SolverStatusEnum status, Optional<Double> solution, SolverType solverType) {
+public record SolverResponse(Optional<String> solver, SolverStatusEnum status, Optional<BigDecimal> solution, SolverType solverType) {
 
     @Override
     public String toString() {
